moved sprint logic into controllers folder
